python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zerolatency"
version = "0.1.0"
description = "Drop-in memory wrappers for Anthropic, OpenAI, and Gemini clients"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "0Latency Team", email = "support@0latency.ai"}
]
keywords = ["ai", "llm", "memory", "anthropic", "openai", "gemini", "claude", "gpt"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "requests>=2.25.0",
    "anthropic>=0.18.0",
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
]

[project.urls]
Homepage = "https://0latency.ai"
Documentation = "https://docs.0latency.ai"
Repository = "https://github.com/0latency/zerolatency-py"
"Bug Reports" = "https://github.com/0latency/zerolatency-py/issues"
